requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
pytest-django==4.8.0
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
